requirements.txt
# Reddit API
praw==7.7.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Web Framework (for future API/Dashboard)
fastapi==0.104.1
uvicorn==0.24.0

# HTTP Requests (for notifications)
requests==2.31.0

# Firebase Cloud Messaging
firebase-admin==6.1.0

# Environment Variables
python-dotenv==1.0.0

# Timezone handling
pytz==2023.3

# Scheduling
apscheduler==3.10.4

# Configuration
pyyaml==6.0.1

# Logging
colorlog==6.7.0

# Date/Time handling
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1